Extract `printValueCommaGroup`
